# tools/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Pappaliiga Stats API",
                 "description":  "REST API for CS2 tournament statistics",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/api/players/{player_id}":  {
                                                   "get":  {
                                                               "tags":  [
                                                                            "players"
                                                                        ],
                                                               "summary":  "Get Player Info",
                                                               "description":  "Get basic player information.",
                                                               "operationId":  "get_player_info_api_players__player_id__get",
                                                               "parameters":  [
                                                                                  "@{name=player_id; in=path; required=True; schema=}"
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  "@{description=Successful Response; content=}",
                                                                                 "422":  "@{description=Validation Error; content=}"
                                                                             }
                                                           }
                                               },
                  "/api/players/{player_id}/seasons":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "players"
                                                                                ],
                                                                       "summary":  "Get Player Season Stats",
                                                                       "description":  "Get player stats across all seasons/divisions.",
                                                                       "operationId":  "get_player_season_stats_api_players__player_id__seasons_get",
                                                                       "parameters":  [
                                                                                          "@{name=player_id; in=path; required=True; schema=}"
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  "@{description=Successful Response; content=}",
                                                                                         "422":  "@{description=Validation Error; content=}"
                                                                                     }
                                                                   }
                                                       },
                  "/api/players/{player_id}/map-stats/{championship_id}":  {
                                                                               "get":  {
                                                                                           "tags":  [
                                                                                                        "players"
                                                                                                    ],
                                                                                           "summary":  "Get Player Map Stats",
                                                                                           "description":  "Get detailed per-map stats with deltas for a player in a specific championship.",
                                                                                           "operationId":  "get_player_map_stats_api_players__player_id__map_stats__championship_id__get",
                                                                                           "parameters":  [
                                                                                                              "@{name=player_id; in=path; required=True; schema=}",
                                                                                                              "@{name=championship_id; in=path; required=True; schema=}"
                                                                                                          ],
                                                                                           "responses":  {
                                                                                                             "200":  "@{description=Successful Response; content=}",
                                                                                                             "422":  "@{description=Validation Error; content=}"
                                                                                                         }
                                                                                       }
                                                                           },
                  "/api/players/":  {
                                        "get":  {
                                                    "tags":  [
                                                                 "players"
                                                             ],
                                                    "summary":  "List Players",
                                                    "description":  "List all players, optionally filtered by season/division/team.",
                                                    "operationId":  "list_players_api_players__get",
                                                    "parameters":  [
                                                                       "@{name=season; in=query; required=False; schema=; description=Filter by season}",
                                                                       "@{name=division; in=query; required=False; schema=; description=Filter by division}",
                                                                       "@{name=team_id; in=query; required=False; schema=; description=Filter by team}",
                                                                       "@{name=limit; in=query; required=False; schema=}"
                                                                   ],
                                                    "responses":  {
                                                                      "200":  "@{description=Successful Response; content=}",
                                                                      "422":  "@{description=Validation Error; content=}"
                                                                  }
                                                }
                                    },
...
              },
...
}